src/requirements.txt
Flask
SQLAlchemy
gunicorn

PyMySQL
coloredlogs
python-dotenv
